provide config via webpack's DefinePlugin.
