+++ .gitignore
@@ -6,7 +6,6 @@
 /build
 node_modules/
 app/src/main/assets/www/
-www_src/config.js
 .idea
 webmaker-*.iml
 npm-debug.log*
