+++ package.json
@@ -5,14 +5,13 @@
   "main": "index.js",
   "scripts": {
     "start": "npm run build && npm-run-all --parallel webserver watch:**",
-    "config": "node ./npm_tasks/build-config.js > www_src/config.js",
     "build": "npm-run-all build:**",
     "build:clean": "rimraf app/src/main/assets/www/ && mkdirp app/src/main/assets/www/",
     "build:static": "mkdirp www_src/static/ && mkdirp app/src/main/assets/ && ncp www_src/static/ app/src/main/assets/www/",
-    "build:js": "npm run config && webpack",
+    "build:js": "webpack",
     "build:css": "npm run watch:css -- --no-watch",
     "build:html": "node ./npm_tasks/bin/build-html",
-    "test": "npm run config && npm-run-all test:**",
+    "test": "npm-run-all test:**",
     "test:mocha": "mocha -R spec --compilers js:babel/register \"./{www_src,npm_tasks}/**/*.test.js\"",
     "test:lint": "jsxhint --harmony \"www_src/{components,html,lib,pages,static}/{**,.}/*.{js,jsx}\"",
     "watch:static": "watch \"npm run build:static\" www_src/static",
